Refactor agent file structure
