+++ languageassistant/agents/conversation/__init__.py
@@ -1,8 +1,7 @@
 """Target language conversation agent"""
 
-from languageassistant.agents.conversation.base import ConversationAgent
-from languageassistant.agents.conversation.conversation_chain import (
-    load_conversation_agent,
-)
+from languageassistant.agents.conversation.agent import ConversationAgent
+from languageassistant.agents.conversation.base import BaseConversationAgent
+from languageassistant.agents.conversation.loader import load_conversation_agent
 
-__all__ = ["ConversationAgent", "load_conversation_agent"]
+__all__ = ["BaseConversationAgent", "ConversationAgent", "load_conversation_agent"]
