+++ languageassistant/agents/planner/__init__.py
@@ -2,6 +2,20 @@
 Lesson planning agent.
 Modified version of langchain chat_planner
 """
-from languageassistant.agents.planner.planner_chain import load_lesson_planner
+from languageassistant.agents.planner.agent import LessonPlannerAgent
+from languageassistant.agents.planner.base import BasePlannerAgent
+from languageassistant.agents.planner.loader import load_lesson_planner
+from languageassistant.agents.planner.schema import (
+    BaseLessonOutputParser,
+    Lesson,
+    LessonOutputParser,
+)
 
-__all__ = ["load_lesson_planner"]
+__all__ = [
+    "BaseLessonOutputParser",
+    "BasePlannerAgent",
+    "LessonOutputParser",
+    "LessonPlannerAgent",
+    "Lesson",
+    "load_lesson_planner",
+]
