+++ languageassistant/agents/planner/base.py
@@ -1,25 +1,16 @@
-from abc import abstractmethod
-from typing import Any, List, Optional
+"""Base abstract planner classes"""
+from abc import ABC, abstractmethod
+from typing import Any
 
 from langchain.callbacks.manager import Callbacks
-from langchain.chains.llm import LLMChain
 from pydantic import BaseModel
 
-from languageassistant.agents.planner.schema import Lesson, LessonOutputParser
+from languageassistant.agents.planner.schema import Lesson
 
 
-class BasePlanner(BaseModel):
+class BasePlannerAgent(BaseModel, ABC):
+    """Abstract base planner agent class"""
+
     @abstractmethod
     def plan(self, inputs: dict, callbacks: Callbacks = None, **kwargs: Any) -> Lesson:
         """Given input, decided what to do."""
-
-
-class LessonPlanner(BasePlanner):
-    llm_chain: LLMChain
-    output_parser: LessonOutputParser
-    stop: Optional[List] = None
-
-    def plan(self, inputs: dict, callbacks: Callbacks = None, **kwargs: Any) -> Lesson:
-        """Given input, decided what to do."""
-        llm_response = self.llm_chain.run(**inputs, stop=self.stop, callbacks=callbacks)
-        return self.output_parser.parse(llm_response)
